+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoAsset.java
@@ -34,6 +34,7 @@ public abstract class VideoAsset {
    * @param remoteUrl remote asset, i.e. typically beginning with {@code https://} or similar.
    * @param streamingFormat which streaming format, provided as a hint if able.
    * @param httpHeaders HTTP headers to set for a request.
+   * @param widevineDrm Widevine DRM configuration, or null for unprotected content.
    * @return the asset.
    */
   @NonNull
@@ -41,8 +42,10 @@ public abstract class VideoAsset {
       @Nullable String remoteUrl,
       @NonNull StreamingFormat streamingFormat,
       @NonNull Map<String, String> httpHeaders,
-      @Nullable String userAgent) {
-    return new HttpVideoAsset(remoteUrl, streamingFormat, new HashMap<>(httpHeaders), userAgent);
+      @Nullable String userAgent,
+      @Nullable WidevineDrmConfiguration widevineDrm) {
+    return new HttpVideoAsset(
+        remoteUrl, streamingFormat, new HashMap<>(httpHeaders), userAgent, widevineDrm);
   }
 
   /**